feat(datastore): get execution payload from archive
Proposers ask relays for block submissions even if the relay never
offered the block. If the relay also never felt the block submission was
the best bid, the relay won't have it. Because non-top block submissions
still get archived in many cases, we have an block submission service
which takes the ones which are safe to propose and makes them available
in Redis. This commit enables the proposer API to read those block
submissions.